requirements.txt
pandas
pytest
pyarrow
pyyaml
pydantic
flake8